% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_quantile_aggregation.R
\name{adaptive_quantile_aggregation}
\alias{adaptive_quantile_aggregation}
\title{Function to apply adaptive version of the quantile aggregation method, Meinshausen et al. (2008)}
\usage{
adaptive_quantile_aggregation(pvals, gamma_min = 0.05)
}
\arguments{
\item{pvals}{a n_bootstrap by n_covariates matrix of the feature statistics, AGCs}

\item{gamma_min}{the pre-specified percentile value used for aggregation, range from 0 to 1}
}
\description{
Function to apply adaptive version of the quantile aggregation method, Meinshausen et al. (2008)
}
