% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhy_threshold.R
\name{bhy_threshold}
\alias{bhy_threshold}
\title{Function to apply the Benjamini-Hochberg-Yekutieli procedure for controlling FDR}
\usage{
bhy_threshold(pvals, reshaping_function = NULL, fdr = 0.1)
}
\arguments{
\item{pvals}{a vector of the feature statistics, AGCs}

\item{reshaping_function}{the default value for reshaping function can be referred by Benjamini & Yekutieli (2001) otherwise, the reshaping function can be referred by Ramdas et al (2017)}

\item{fdr}{the pre-specified false discovery rate level}
}
\value{
The threshold value corresponding tthe Benjamini-Hochberg-Yekutieli procedure
}
\description{
Function to apply the Benjamini-Hochberg-Yekutieli procedure for controlling FDR
}
