% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_L1.R
\name{build_L1}
\alias{build_L1}
\title{Funtion to construct the Laplacian matrix based on the graphical structure}
\usage{
build_L1(L)
}
\arguments{
\item{L}{p by p (weighted) adjacency matrix of a graph.}
}
\description{
Funtion to construct the Laplacian matrix based on the graphical structure
}
