% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvGrace.R
\name{cvGrace}
\alias{cvGrace}
\title{Function to perform cross-validation for Grace from 'Grace' r package}
\usage{
cvGrace(X, Y, L, lambda.L, lambda.1, lambda.2, K = 10)
}
\arguments{
\item{X}{standardized n (number of rows) by p (number of columns) design matrix.}

\item{Y}{n by 1 vector of the response variable.}

\item{L}{p by p symmetric matrix of the penalty weight matrix.}

\item{lambda.L}{tuning parameters of the penalty weight matrix.}

\item{lambda.1}{tuning parameters of the L_1 penalty.}

\item{lambda.2}{tuning parameters of the ridge penalty.}

\item{K}{number of folds in cross-validation.}
}
\description{
Function to perform cross-validation for Grace from 'Grace' r package
}
