% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_pval.R
\name{empirical_pval}
\alias{empirical_pval}
\title{Function of intermediary products derived from the conversion of feature statistics (single knockoffs) to aggregated feature statistics.}
\usage{
empirical_pval(test_score, offset = 1)
}
\arguments{
\item{test_score}{a vector of the feature statistics obatianed from single knockoff procedure}

\item{offset}{a modified value of the conversion procedure}
}
\value{
a vector of the intermediary products
}
\description{
Function of intermediary products derived from the conversion of feature statistics (single knockoffs) to aggregated feature statistics.
}
