% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_threshold.R
\name{fdr_threshold}
\alias{fdr_threshold}
\title{Function to compute the corresponding threshold (BY/BH) for controlling FDR}
\usage{
fdr_threshold(pvals, fdr = 0.1, method = "bhq", reshaping_function = NULL)
}
\arguments{
\item{pvals}{a vector of the feature statistics, AGCs}

\item{fdr}{the pre-specified false discovery rate level}

\item{method}{the test procedure applied to compute the threshold. The defult method to compute threshold is the the Standard Benjamini-Hochberg procedure.}

\item{reshaping_function}{the default value for reshaping function can be referred by Benjamini & Yekutieli (2001) otherwise, the reshaping function can be referred by Ramdas et al (2017)}
}
\value{
The threshold value corresponding to the test procedure
}
\description{
Function to compute the corresponding threshold (BY/BH) for controlling FDR
}
