% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_quantile_aggregation.R
\name{fixed_quantile_aggregation}
\alias{fixed_quantile_aggregation}
\title{Function to apply quantile aggregation function based on Meinshausen et al (2008)}
\usage{
fixed_quantile_aggregation(pvals, gamma = 0.3)
}
\arguments{
\item{pvals}{a n_bootstrap by n_covariates matrix of the feature statistics, AGCs}

\item{gamma}{the pre-specified percentile value used for aggregation, range from 0 to 1}
}
\description{
Function to apply quantile aggregation function based on Meinshausen et al (2008)
}
