% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grace_multi_knockoff.R
\name{Grace_multi_knockoff}
\alias{Grace_multi_knockoff}
\title{Function to conduct Grace-AKO and Grace}
\usage{
Grace_multi_knockoff(
  X,
  Y,
  L_1,
  L,
  fdr,
  lambda.L,
  lambda.1,
  lambda.2,
  n_bootstraps = 25,
  gamma = 0.3,
  gamma_min = 0.1
)
}
\arguments{
\item{X}{standardized n (number of rows) by p (number of columns) design matrix.}

\item{Y}{centered n by 1 vector of the response variable.}

\item{L_1}{p by p (weighted) processed matrix of a graph for Grace-AKO.}

\item{L}{p by p (weighted) adjacency matrix of a graph for Grace.}

\item{fdr}{the pre-specified false discovery rate level}

\item{lambda.L}{tuning parameters of the penalty weight matrix}

\item{lambda.1}{tuning parameters of the L_1 penalty}

\item{lambda.2}{tuning parameters of the ridge penalty}

\item{n_bootstraps}{the number of multiple knockoffs}

\item{gamma}{the pre-specified percentile value used for fixed aggregation}

\item{gamma_min}{the pre-specified percentile value used for adaptive aggregation}
}
\description{
Function to conduct Grace-AKO and Grace
}
