% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdr_tpp_base.R
\name{mfdr_tpp_base}
\alias{mfdr_tpp_base}
\title{Function to compute the modified false discovery rate and true positive proportion for baseline model (e.g., Grace)}
\usage{
mfdr_tpp_base(q, beta_true, beta_est)
}
\arguments{
\item{q}{the pre-specified false discovery rate level}

\item{beta_true}{the true coefficient of graph-structured covariates}

\item{beta_est}{the estimated coefficient of graph-structured covariates}
}
\description{
Function to compute the modified false discovery rate and true positive proportion for baseline model (e.g., Grace)
}
