% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdr_tpp_knockoff.R
\name{mfdr_tpp_knockoff}
\alias{mfdr_tpp_knockoff}
\title{Function to compute the modified false discovery rate and true positive proportion for Grace-AKO}
\usage{
mfdr_tpp_knockoff(q, w, t, beta_true, beta_est)
}
\arguments{
\item{q}{the pre-specified false discovery rate level}

\item{w}{a 1 by n_covariates vector of aggregated feature statistics}

\item{t}{a data-dependent threshold}

\item{beta_true}{the true coefficient of graph-structured covariates}

\item{beta_est}{the estimated coefficient of graph-structured covariates}
}
\description{
Function to compute the modified false discovery rate and true positive proportion for Grace-AKO
}
