% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_aggregation.R
\name{quantile_aggregation}
\alias{quantile_aggregation}
\title{Function to apply the corresponding quantile aggregation function}
\usage{
quantile_aggregation(pvals, gamma = 0.5, gamma_min = 1e-05, adaptive = FALSE)
}
\arguments{
\item{pvals}{a n_bootstrap by n_covariates matrix of the feature statistics, AGCs}

\item{gamma}{the pre-specified percentile value used for the fixed quantile aggregation, range from 0 to 1}

\item{gamma_min}{the pre-specified percentile value used for the adaptive quantile aggregation, range from 0 to 1}

\item{adaptive}{the statement to determine the quantile aggregation function}
}
\description{
Function to apply the corresponding quantile aggregation function
}
